mlrunnerclient/requirements.txt
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
numpy==1.19.1
PyYAML==5.3.1
requests==2.24.0
urllib3==1.25.10
